+++ platforms/azure/giantnetes/variables.tf
@@ -110,7 +110,7 @@ variable "flatcar_linux_channel" {
 variable "flatcar_linux_version" {
   description = "Flatcar linux version."
   type        = string
-  default     = "2345.3.1"
+  default     = "2512.2.1"
 }
 
 variable "vault_image_publisher" {