+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2020-08-05
+
 ### Changed
 
 - Update Flatcar Linux version to `2512.2.1`.
@@ -55,7 +57,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add github release workflows.
 
-[Unreleased]: https://github.com/giantswarm/giantnetes-terraform/compare/v1.0.5...HEAD
+[Unreleased]: https://github.com/giantswarm/giantnetes-terraform/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/giantswarm/giantnetes-terraform/compare/v1.0.5...v1.1.0
 [1.0.5]: https://github.com/giantswarm/giantnetes-terraform/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/giantswarm/giantnetes-terraform/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/giantswarm/giantnetes-terraform/compare/v1.0.2...v1.0.3